refine failpoints test
